+++ source/mechcmd2/debugging.cpp
@@ -19,7 +19,7 @@
 #include <GameOS\ToolOS.hpp>
 
 HGOSFONT3D GameDebugWindow::font = NULL;
-long GameDebugWindow::fontHeight = 0;
+int32_t GameDebugWindow::fontHeight = 0;
 
 //***************************************************************************
 //	GAME DEBUG WINDOW class
@@ -35,7 +35,7 @@ void GameDebugWindow::setFont (PSTR fontFile) {
 		font = gos_LoadFont(fontFile);
 		gos_TextSetAttributes(font, 0xffffffff, 1.0, true, true, false, false);
 	}
-	DWORD height, width;
+	ULONG height, width;
 	gos_TextStringLength(&width, &height, "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ");
 	fontHeight = height;
 }
@@ -61,8 +61,8 @@ void GameDebugWindow::render (void) {
 	if (!display)
 		return;
 
-	long curY = pos[1] + 5;
-	for (long i = linePos; i < MAX_DEBUG_WINDOW_LINES; i++) {
+	int32_t curY = pos[1] + 5;
+	for (int32_t i = linePos; i < MAX_DEBUG_WINDOW_LINES; i++) {
 		gos_TextSetPosition(pos[0] + 5, curY);
 		curY += fontHeight;
 		gos_TextDraw(textBuffer[i]);
